+++ chain-crypto/src/ec/p256k1.rs
@@ -445,3 +445,5 @@ mod test {
         assert_eq!(element, element2);
     }
 }
+
+gen_group_tests!();
