+++ chain-crypto/src/ec/ristretto255.rs
@@ -382,3 +382,5 @@ mod test {
         assert_eq!(element, element2)
     }
 }
+
+gen_group_tests!();
